nixos/postgresql: add `lc_{collate,ctype}` options to `ensureDatabases`
Since `matrix-synapse` v1.10.0[1], the service requires a database with
the values LC_{COLLATE,CTYPE} set to `C`[2]. To quote the postgresql
docs[3]:

> If you want the system to behave as if it had no locale support,
> use the special locale name C, or equivalently POSIX.

Existing databases will continue to work with a warning, but this is a
hard requirement for fresh setups (like in our VM-test).

This patch enhances the `ensureDatabases` option to make it possible to
pass an attribute set with those settings or a string to create a new
database.

[1] https://github.com/matrix-org/synapse/releases/tag/v1.10.0
[2] https://github.com/matrix-org/synapse/blob/v1.10.0/docs/postgres.md#set-up-database
[3] https://www.postgresql.org/docs/12/locale.html