subsurface: 5.0.10 -> 6.0.5231
subsurface changed from tagged releases to nightly builds
from which a somewhat stable release is marked as a "weekly release"
which is published on subsurface-divelog.org
Currently there is no tag on the main repository pointing
to the commit that was used to create the nightly build
that was finally selected as the "weekly release".
Once these tags are published, an `updateScript` can be
put into place to track them. Until then this is a manual
process.

Upstream Documentation:
- https://github.com/subsurface/subsurface/releases/tag/v4.9.4
- https://github.com/subsurface/nightly-builds/releases
- https://subsurface-divelog.org/current-release/